+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/MessageEntity/textsources/RegularTextSource.kt
@@ -12,9 +12,9 @@ import dev.inmo.tgbotapi.utils.internal.regularMarkdownV2
 data class RegularTextSource @RiskFeature(DirectInvocationOfTextSourceConstructor) constructor (
     override val source: String
 ) : TextSource {
-    override val asMarkdownSource: String by lazy { source.regularMarkdown() }
-    override val asMarkdownV2Source: String by lazy { source.regularMarkdownV2() }
-    override val asHtmlSource: String by lazy { source.regularHtml() }
+    override val markdown: String by lazy { source.regularMarkdown() }
+    override val markdownV2: String by lazy { source.regularMarkdownV2() }
+    override val html: String by lazy { source.regularHtml() }
 }
 
 @Suppress("NOTHING_TO_INLINE")