+++ tgbotapi.extensions.utils/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/utils/formatting/LinksFormatting.kt
@@ -4,7 +4,6 @@ import dev.inmo.tgbotapi.types.*
 import dev.inmo.tgbotapi.types.MessageEntity.textsources.link
 import dev.inmo.tgbotapi.types.ParseMode.*
 import dev.inmo.tgbotapi.types.chat.abstracts.*
-import dev.inmo.tgbotapi.types.chat.abstracts.extended.ExtendedChat
 import dev.inmo.tgbotapi.types.chat.abstracts.extended.ExtendedPublicChat
 import dev.inmo.tgbotapi.types.message.abstracts.Message
 
@@ -92,17 +91,17 @@ val StickerSetName.stickerSetLink
 @Deprecated("Use extension `stickerSetLink` + getting of `asMarkdownV2Source` property")
 fun makeLinkToAddStickerSetInMarkdownV2(
     stickerSetName: StickerSetName
-) = stickerSetName.stickerSetLink.asMarkdownV2Source
+) = stickerSetName.stickerSetLink.markdownV2
 /**
  * @return Link for adding of sticker set with name [stickerSetName] with formatting for [Markdown]
  */
 @Deprecated("Use extension `stickerSetLink` + getting of `asMarkdownSource` property")
-fun makeLinkToAddStickerSetInMarkdown(stickerSetName: StickerSetName) = stickerSetName.stickerSetLink.asMarkdownSource
+fun makeLinkToAddStickerSetInMarkdown(stickerSetName: StickerSetName) = stickerSetName.stickerSetLink.markdown
 /**
  * @return Link for adding of sticker set with name [stickerSetName] with formatting for [HTML]
  */
 @Deprecated("Use extension `stickerSetLink` + getting of `asHtmlSource` property")
-fun makeLinkToAddStickerSetInHtml(stickerSetName: StickerSetName) = stickerSetName.stickerSetLink.asHtmlSource
+fun makeLinkToAddStickerSetInHtml(stickerSetName: StickerSetName) = stickerSetName.stickerSetLink.html
 /**
  * Create a link for adding of sticker set with name [stickerSetName]. Was added thanks to user Djaler and based on
  * https://github.com/Djaler/evil-bot/blob/master/src/main/kotlin/com/github/djaler/evilbot/utils/StickerUtils.kt#L6-L8